Securing API with various authentication methods

+++ Restuarants.Application/Restuarants/Commands/DeleteRestuarant/DeleteRestuarantCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restuarants.Domain.Abstracts;
+using Restuarants.Domain.Exceptions;
 using Restuarants.Domain.Repositories;
 
 namespace Restuarants.Application.Restuarants.Commands.DeleteRestuarant
@@ -8,11 +10,13 @@ namespace Restuarants.Application.Restuarants.Commands.DeleteRestuarant
     {
         private readonly IRestuarantsRepository _restuarantsRepository;
         private readonly ILogger<DeleteRestuarantCommandHandler> _logger;
+        private readonly IRestuarantAuthorizationService _restuarantAuthorizationService;
 
-        public DeleteRestuarantCommandHandler(ILogger<DeleteRestuarantCommandHandler> logger, IRestuarantsRepository restuarantsRepository)
+        public DeleteRestuarantCommandHandler(ILogger<DeleteRestuarantCommandHandler> logger, IRestuarantsRepository restuarantsRepository, IRestuarantAuthorizationService restuarantAuthorizationService)
         {
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _restuarantsRepository = restuarantsRepository ?? throw new ArgumentNullException(nameof(restuarantsRepository));
+            _restuarantAuthorizationService = restuarantAuthorizationService ?? throw new ArgumentNullException(nameof(restuarantAuthorizationService));
         }
 
         public async Task<bool> Handle(DeleteRestuarantCommand command, CancellationToken cancellationToken)
@@ -23,7 +27,12 @@ namespace Restuarants.Application.Restuarants.Commands.DeleteRestuarant
 
             if(restuarant is null)
             {
-                throw new ApplicationException($"Restuarant with {command.Id} does not exist, please try again");
+                throw new NotFoundException($"Restuarant with {command.Id} does not exist, please try again");
+            }
+
+            if(_restuarantAuthorizationService.Authorize(restuarant, Domain.Constants.ResourceOperation.Delete) == false)
+            {
+                throw new ForbidException();
             }
 
             await _restuarantsRepository.DeleteRestuarantAsync(command.Id);
